move from regex::Regex to regex::bytes::Regex
